Resolve paths relative to servlet (webapp) context
When dealing with MASTERKEY_ROOT_CONFIG_DIR.

This allows to package configuration along with the webapp or find it easily
when running the webapp exploded. This is particularly useful for situations
where app server working dir cannot be predicted (and thus relative paths are
useless) -- e.g integration testing or development.

Webapp relative path can be specified as:

war://path/within/the/webapp